Refactor: EnteredChatCard -> ImageChatCard 리팩토링 #39
