--- components/ui/Card/ImageChatCard.tsx
@@ -1,11 +1,0 @@
-import ChatCard from '.';
-
-export default function ImageChatCard() {
-  return (
-    <ChatCard>
-      <ChatCard.Header>
-        <div>d</div>
-      </ChatCard.Header>
-    </ChatCard>
-  );
-}
