--- components/ui/ChatCard/ChatCardInfo.tsx
@@ -1,24 +1,0 @@
-import Avatar from '../Avatar';
-import Divider from '../Divider';
-import FlexBox from '../FlexBox';
-
-interface ChatCardInfoProps {
-  masterUserImg: string;
-  masterUserName: string;
-  participants: number;
-}
-
-export default function ChatCardInfo({
-  participants,
-  masterUserImg,
-  masterUserName,
-}: ChatCardInfoProps) {
-  return (
-    <FlexBox className="text-[#74787D] gap-2 text-sm tablet:text-base">
-      <Avatar image={masterUserImg} name={masterUserName} />
-      <p className="font-medium">{masterUserName}</p>
-      <Divider type="vertical" />
-      <p>{participants}명</p>
-    </FlexBox>
-  );
-}
