+++ components/ui/ChatCard/NormalChatCard.tsx
